requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.5
emoji==2.10.1
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
halo==0.0.31
httplib2==0.22.0
idna==3.6
log-symbols==0.0.14
oauthlib==3.2.2
prettytable==3.10.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
spinners==0.0.24
termcolor==2.4.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13